+++ examples/terraform/backend.tf
@@ -1,0 +1,7 @@
+terraform {
+  backend "s3" {
+    bucket = "tanelso2-terraform-state"
+    key    = "this.state"
+    region = "us-east-2"
+  }
+}