+++ test/hack_a_lisp/core_test.clj
@@ -1,7 +1,0 @@
-(ns hack-a-lisp.core-test
-  (:require [clojure.test :refer :all]
-            [hack-a-lisp.core :refer :all]))
-
-(deftest a-test
-  (testing "FIXME, I fail."
-    (is (= 0 1))))