+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+* Added evaluator for Terraform
 
 ## [0.1.1] — 2024-08-07
 * Add basic cli functions